\name{atime_versions}
\alias{atime_versions}
\title{Asymptotic timing of git versions}
\description{Computation time and memory for a single R expression
  evaluated using several different git versions.}
\usage{
atime_versions(
 pkg.path, N, setup, expr, sha.vec=NULL,
 times=10, seconds.limit=0.01, verbose=FALSE,
 pkg.edit.fun=pkg.edit.default, result=FALSE,
 N.env.parent=NULL, setup.version=NULL,
 checkout.path=pkg.path,
 ...)
}
\arguments{
  \item{pkg.path}{Path to R package in a git repo.}
  \item{N}{numeric vector of data sizes to vary.}
  \item{setup}{
    expression to evaluate for every data size, before timings.
    If one of the versions contains name HEAD, then
    \code{Package::} is replaced with
    \code{Package.SHA_for_HEAD::}
    (in other words, setup is run using the HEAD commit).
  }
  \item{expr}{
    code with package double-colon prefix, for example
    \code{PKG::fun(argA, argB)}, where PKG is the name of the
    package specified by \code{pkg.path}. This code will be
    evaluated for each different package version, by replacing
    \code{PKG::} by \code{PKG.SHA::}.
    To run different versions of implicitly-called functions
    like \code{DT[i,j]}, you need
    to call them explicitly, as in 
    \code{data.table:::`[.data.table`(DT,i,j)}.
  }
  \item{sha.vec}{named character vector / list of versions.}
  \item{times}{number of times to evaluate each timed expression.}
  \item{seconds.limit}{if the median timing of any expression exceeds
    this many seconds, then no timings for larger N are computed.}
  \item{verbose}{logical, print messages after every data size?}
  \item{pkg.edit.fun}{function called to edit package before
    installation, should typically replace instances of \code{PKG} with
    \code{PKG.SHA},
    default works with Rcpp packages that have \code{useDynLib(pkg)}
    in NAMESPACE (no \code{.registration=TRUE}).}
  \item{result}{logical or function, passed to \code{\link{atime}}.}
  \item{N.env.parent}{environment to use as parent of environment
  created for each data size N, or NULL to use default parent env.}
  \item{setup.version}{expression to evaluate for each version, prior to
  timings.}
  \item{checkout.path}{path to repo in which we run git checkout prior
    to package installation. Default is \code{pkg.path}, but can be set
    to a sub-directory with a git submodule. }
  \item{\dots}{named versions.}
}
\details{
  For convenience, versions can
  be specified either as code (\dots), data (\code{sha.vec}), or both.
  Each version should be either \code{""} (to use currently installed
  version of package, or if missing, install most recent
  version from CRAN) or a SHA1 hash, which is passed as branch
  to git checkout; version names used to identify/interpret
  the output/plots.
}
\value{list of class atime with elements \code{seconds.limit} (numeric
  input param), \code{timings} (data table of results).}
\seealso{
  \code{\link{atime_versions_exprs}} 
  converts \code{expr} into a list of expressions, one for each version,
  passed to \code{\link{atime}} as the \code{expr.list} argument.
}

\author{Toby Dylan Hocking}

\examples{

if(FALSE){

  tdir <- tempfile()
  dir.create(tdir)
  gert::git_clone("https://github.com/tdhock/binsegRcpp", tdir)
  atime.list <- atime::atime_versions(
    pkg.path=tdir,
    setup={
      max.segs <- as.integer(N/2)
      data.vec <- 1:N
    },
    expr=binsegRcpp::binseg_normal(data.vec, max.segs),
    cv="908b77c411bc7f4fcbcf53759245e738ae724c3e",
    "rm unord map"="dcd0808f52b0b9858352106cc7852e36d7f5b15d",
    mvl_construct="5942af606641428315b0e63c7da331c4cd44c091")
  plot(atime.list)

  atime::atime_versions_remove("binsegRcpp")

  gdir <- tempfile()
  dir.create(gdir)
  gert::git_clone("https://github.com/tdhock/grates", gdir)
  glist <- atime::atime_versions(
    file.path(gdir,"pkg"),
    current = "1aae646888dcedb128c9076d9bd53fcb4075dcda",
    old     = "51056b9c4363797023da4572bde07e345ce57d9c",
    setup   = date_vec <- rep(Sys.Date(), N),
    expr    = grates::as_yearmonth(date_vec))
  plot(glist)

  atime::atime_versions_remove("grates")

}

}
