\name{issue109}
\alias{issue109}
\docType{data}
\title{
  atime results for issue 109
}
\description{
  atime was run on GitHub Actions, to test data.table.
  One test had a speed improvement, and was shown on the right of the
  facet plot.
}
\usage{data("issue109")}
\format{List from running \code{atime_pkg()}.}
\references{
  https://github.com/Rdatatable/data.table/pull/7401#issuecomment-3457752439
}
\examples{

data(issue109, package="atime")

}
\keyword{datasets}
