\name{glob_find_replace}
\alias{glob_find_replace}
\title{Find and replace within files}
\description{Find and replace for every file specified by glob.}
\usage{glob_find_replace(glob, FIND, REPLACE)}
\arguments{
  \item{glob}{character string: glob defining files.}
  \item{FIND}{character string: regex to find.}
  \item{REPLACE}{character string: regex to use for replacement.}
}
\value{nothing.}

\author{Toby Dylan Hocking}

\examples{

## see vignette("data.table", package="atime")

}
